fix user id integer type size

+++ src/migration/1520195917088-Users.ts
@@ -5,7 +5,7 @@ export class User1520195917088 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query(`
       CREATE TABLE users (
-        id integer PRIMARY KEY,
+        id bigint PRIMARY KEY,
         username text,
         prestige integer NOT NULL
       );
